+++ hois_back/src/main/java/ee/hitsa/ois/web/dto/curriculum/CurriculumVersionOccupationModuleYearCapacityDto.java
@@ -1,5 +1,7 @@
 package ee.hitsa.ois.web.dto.curriculum;
 
+import java.math.BigDecimal;
+
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 
@@ -7,16 +9,14 @@ import ee.hitsa.ois.domain.curriculum.CurriculumVersionOccupationModuleYearCapac
 import ee.hitsa.ois.util.EntityUtil;
 import ee.hitsa.ois.web.commandobject.VersionedCommand;
 
-public class CurriculumVersionOccupationModuleYearCapacityDto  extends VersionedCommand {
+public class CurriculumVersionOccupationModuleYearCapacityDto extends VersionedCommand {
 
     private Long id;
-
     @NotNull
     @Min(1)
     private Integer studyYearNumber;
-
     @NotNull
-    private Double credits;
+    private BigDecimal credits;
 
     public static CurriculumVersionOccupationModuleYearCapacityDto of(CurriculumVersionOccupationModuleYearCapacity capacity) {
         CurriculumVersionOccupationModuleYearCapacityDto dto =
@@ -40,12 +40,11 @@ public class CurriculumVersionOccupationModuleYearCapacityDto  extends Versioned
         this.studyYearNumber = studyYearNumber;
     }
 
-    public Double getCredits() {
+    public BigDecimal getCredits() {
         return credits;
     }
 
-    public void setCredits(Double credits) {
+    public void setCredits(BigDecimal credits) {
         this.credits = credits;
     }
-
 }
